add ecadd precompile invalid inputs
